requirements.txt
alembic==1.5.7
autopep8==1.5.6
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-CKEditor==0.4.4.1
Flask-FontAwesome==0.1.5
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
migrate==0.3.8
nose2==0.10.0
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
